Implemented generate_tile_coordinates.
Also changed the default behavior when the cropping reaches the right/bottom of the image. It now uses the default of Image.crop() to pad the right and bottom with black rather than making it 'hit' the right/bottom and go no further.

Might change this later, but it will make matching bounding boxes easier later.