Installing external dependencies for Sphinx workflow.
I believe the error is due to Sphinx not being able to cleanly import external dependencies (right now only Pillow), so it fails to import tiling.py as a module.
Added a line that will (read should) install the external requirements. This will add overhead when we eventually add Yolov8, but I have been made aware of a module called mock that will allow us to mock an import so we don't have to get GitHub to download it.