Fixed bug in tile_annotations.
There exists a bug in tile_annotations where the annotation's coordinates were simply not getting corrected based on which tile they belonged to. This would cause a box to be in pixels that didn't exist in the image.

To solve this, there is a new method for BoundingBox that sets the box's values. Under the hood this just returns a new object with the new values (and the same category). Keypoint was updated to have the same function, and although it doesn't strictly need to, it also returns a whole new keypoint. Also added a correct_keypoint method. Note that you can change the bounding box of a keypoint such that it is invalid, but you cannot change its keypoint in a way that makes it invalid.