Add a new method to Keypoint: set_box_and_keypoint
Ran into an issue with set_box where setting a box that didn't contain the keypoint threw an exception. This creates a deadlock situation where the keypoint cannot be moved until the box is, and the box cannot be moved until the keypoint is. This new method does both at once so that this doesn't occur.