Remove setter in favor of immutability.
Since this repo is functional first, setters are a no-go. Instead, a brand new object should be created. I can see how this is a bit tricky here since label is derived from the boxes, but also might need to just be set, so I created an alternative constructor that matches what the other constructor was, and made create_cluster_label a staticmethod so that the alternative constructor doesn't have to need an instance of the class to use it.