Change cluster_boxes to be a higher order function.
cluster_boxes was feeling a little weird with all the optional parameters, plus needing to validate parameters for all possible clustering methods. So, this function is now a higher order function that accepts the other clustering functions as a parameter and passes the key word arguments in as parameters.

Ideally, where this is called the values would be partially applied before passing in, such that no kwargs are passed.

Also, moved validation of parameters to the function that needs them.