Switch to partial application in tests.
Switch to the new use of cluster_boxes that uses a passed in function to do clustering.

I went with the partial application way of doing it to show how it can be done, but it is equally acceptable to pass in the clustering functions directly and supply the needed arguments as kwargs.